// src/api/types/movieVideos200ResultsItem.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */

export type MovieVideos200ResultsItem = {
  iso_639_1?: string;
  iso_3166_1?: string;
  name?: string;
  key?: string;
  site?: string;
  size?: number;
  type?: string;
  official?: boolean;
  published_at?: string;
  id?: string;
};